winchecksec: remove unintentionally public APIs
Closes #1036.
